HW-1-App-init-and-API-set-up
App basic architecture init
