+++ spotify-api-player/src/components/MainScreen/Player/PlayerComponent.js
