+++ spotify-api-player/src/constants/utils.js
@@ -1,0 +1,10 @@
+export const hash = window.location.hash
+  .substring(1)
+  .split("&")
+  .reduce(function (initial, item) {
+    if (item) {
+      var parts = item.split("=");
+      initial[parts[0]] = decodeURIComponent(parts[1]);
+    }
+    return initial;
+  }, {});
